backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.12.0
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
requests==2.31.0
